requirements.txt
# Custom
av2
lightning

# Original
absl-py
astor
beautifulsoup4
brotlipy
bs4
cachetools
certifi
contextlib2
cycler
dgl
filelock
future
google-auth
google-auth-oauthlib
grpcio
h5py
hyperopt
json-tricks
littleutils
Markdown
mkl-fft
mkl-service
munkres
netifaces
nni
oauthlib
ogb
olefile
outdated
pandas
prettytable
protobuf
psutil
pyasn1
pyasn1-modules
PyGSP
pymongo
pyparsing
PythonWebHDFS
PyYAML
requests-oauthlib
responses
rsa
schema
scipy
seaborn
simplejson
sip
soupsieve
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
termcolor
torch-geometric
wcwidth
websockets
Werkzeug
xlrd
tensorboardx
ruamel.yaml
